third_party/blink/web_tests/fast/css/font-face-descending-unicode-range.html

<!doctype html>
<html>
<head>
<title>Test for Bug 41339 - unicode-range property only with a descending range causes a crash</title>
<style>
@font-face {
    font-family: 'myfont';
    src: local(Courier), local('Courier New'), local('Cousine Regular'); /* Use monospace font */
    unicode-range: U+062-60;
}
</style>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="description"></div>
<div style="font-family: 'myfont'">
<span id="i_s" style="border:solid 1px">iiiiiiiiii</span><br/>
<span id="x_s" style="border:solid 1px">XXXXXXXXXX</span><br/>
</div>
<div id="console"></div>
<script>
description("Interval unicode-ranges that descend are invalid and omitted. Without any valid ranges, the descriptor is omitted. As a result, the @font-face rule in this file is applied to all characters. This test passes if the following i's and X's are shown in Courier and they have the same length on screen. As of r62096, opening this file causes a crash instead.");

function validate() {
  shouldBe("document.getElementById('i_s').offsetWidth", "document.getElementById('x_s').offsetWidth");
  finishJSTest();
}

window.jsTestIsAsync = true;
document.fonts.ready.then(validate);
</script>
</body>
</html>
